Feedback_Form-main/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
motor
pymongo>=4.8.0
aiofiles
